moved Attribution to own component
